requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
python-dateutil 